don’t need to pass loading prop through from item page
