+++ catalogue/webapp/pages/item.js
@@ -1,7 +1,5 @@
 // @flow
 import { type Context } from 'next';
-import { useEffect, useState } from 'react';
-import Router from 'next/router';
 import {
   type Work,
   type CatalogueApiError,
@@ -146,22 +144,6 @@ const ItemPage = ({
     linkKey: 'page',
     ...sharedPaginatorProps,
   };
-  const [loading, setLoading] = useState(false);
-  function routeChangeStart(url: string) {
-    setLoading(true);
-  }
-  function routeChangeComplete(url: string) {
-    setLoading(false);
-  }
-  useEffect(() => {
-    Router.events.on('routeChangeStart', routeChangeStart);
-    Router.events.on('routeChangeComplete', routeChangeComplete);
-
-    return () => {
-      Router.events.off('routeChangeStart', routeChangeStart);
-      Router.events.off('routeChangeComplete', routeChangeComplete);
-    };
-  }, []);
 
   return (
     <CataloguePageLayout
@@ -210,7 +192,6 @@ const ItemPage = ({
           canvasIndex={canvasIndex}
           iiifImageLocationUrl={iiifImageLocationUrl}
           imageUrl={imageUrl}
-          loading={loading}
         />
       )}
     </CataloguePageLayout>
